Remove usage of ratelimits on wireproto push
Summary:
This diff removes usages of total_file_changes so that the field can be deprecated in thrift.

In practice the endpoint is no longer used by the client, as per https://fburl.com/scuba/mononoke_test_perf/dv6w679l and previous conversations with the client team, so I've removed the usage of the ratelimiting crate entirely rather than refactoring it to fit with the new config.

Reviewed By: andreacampi

Differential Revision: D68214487

fbshipit-source-id: c76ed5374528f06c4ace84f1977892a2504798d3