Integration test highlighting incorrect capture of refs pointing to trees and blobs
Summary:
For regular `refs` (i.e. `refs` pointing to commits and tags), we store the name of the ref without the `refs/` prefix. However, that is not the case for content refs where we include the prefix in the name. This mismatch is problematic because it results in Mononoke Git server returning the refs as `refs/refs/branch_to_tree` which is incorrect.

This integration test highlights the incorrect behaviour which will be fixed in follow up diff.

Reviewed By: lmvasquezg

Differential Revision: D68265245

fbshipit-source-id: 9db57c8648be43bf60395ea0c0373fe65276f351