Fix handling of content refs storage
Summary:
As mentioned in the previous diff, we do not store refs pointing to trees and blobs the same way as we store other refs. The changes in this diff ensures that we maintain parity with the other ref handling.

Note that this also changes the output of the Mononoke Git test for refs pointing to trees and blobs.

Reviewed By: lmvasquezg

Differential Revision: D68265327

fbshipit-source-id: 559ddf8d1fd1db33c96c7ac53027d1ce0945ddbc