Logger class that includes timestamp
Summary:
We've had a very bare-bones logging system for a long time now, and it's been okay. But it would be nice to have time stamps on lines of the log file to correlate with analytics.

Let's add a simple Logger class instead of an interface. We have 4 current usages which we can have different implementations for:
- stdout (when in dev mode), which can use ANSI colors
- file (typical `sl isl` use)
- vscode output channel (typical vscode use, which is most common)
- integration tests

Note: client side logs current only output in the client window, this diff does not affect client side logging (which is hardly used)

Reviewed By: nsblake

Differential Revision: D68190739

fbshipit-source-id: f6a87877e1b3337f4d31ddb39afa791d9cd57e79