hooks: replace HookStateProvider with HookRepo
Summary:
The HookStateProvider is an abstraction over the repo that dates back to when hooks were written in Lua and needed a well-defined abstraction to access repository information.  Now that hooks are written as normal Rust code within the Mononoke binary, this abstraction just gets in the way of accessing repository data.  Whenever a hook needs a new piece of data, we must add a new abstraction to the state provider just so that the new hook can access it.

Instead, let's give hooks access to an actual repository.  Since the hook manager is a repo facet itself, it needs its own concrete repo type, which we call `HookRepo`.  This means hooks have access to a preset list of repository facets, and we guarantee they will always be available.  Future hooks that need additional facets will need to add them to `HookRepo` to gain access.

This diff is large, as it changes some widely-used trait interfaces.  A summary of the changes is:

- Replace the HookStateProvider trait and RepoHookStateProvider with a new facet container: HookRepo.
- References to `&dyn HookStateProvider` become `&HookRepo`.
- Methods on `HookStateProvider` become inherent methods on `HookRepo`.
- All call-sites are updated.
- `HookRepo` is constructed either using `build_from` the main repo type, or by assembling it directly out of the required facets.

Differential Revision: D68216792

fbshipit-source-id: 1cec7b7db263ba91cd328a9b9e791aba8995e46a