hooks: remove TextOnlyFileContentProvider
Summary:
This additional layer in the hook content/state provider abstration does very little.  Let's remove it and lift its features into the repo implementation (which is the only place it's currently used).

Incidentally, its removal revealed a problem that it was hiding.  Because `get_file_text` elides file contents with nulls in, the check for symlinks containing nulls didn't work in prod, as these values were being elided.  It worked in the test because the test didn't use `TextOnlyFileContentProvider`.  To fix this, I've added `get_file_bytes` that does the same as `get_file_text` but excludes the null-byte check.

Reviewed By: RajivTS

Differential Revision: D68216795

fbshipit-source-id: e73fae2869dfd9981be19eecc327e4581802f739