Add "uncommitted changes" title to simplify hint
Summary:
The idea here is to remove the instruction hint `Diff chunks under "You are here" will be left in the working copy.` by adding an "uncommitted changes" label. This, I think, makes it reasonably clear which changes are going to be left as uncommitted changes.

Simplifying the banner makes it more likely that people actually read all the info there.

While doing this, I think we can also remove the hint about changes under files. I think this is more clear by the padding on changes.

Reviewed By: quark-zju

Differential Revision: D68286979

fbshipit-source-id: d7bb90875c8f52ef162039335641d8e60b1bd9e5