Integration test highlighting lack of support for standalone refs pointing to trees in Mononoke Git read path
Summary:
We now support the standard scenario of refs pointing to trees and blobs in Mononoke Git clones/fetches/pulls. However, if we have a repo with a ref pointing to a tree that is not pointed to by any commit (i.e. standalone tree) then we fail while cloning such a repo. This is because we return just the tree and not the entire contents contained within the directory represented by the tree.

This _can_ be done but it would be expensive not just for this case but for multiple other cases. Since this is an edge case anomaly, we are creating an integration test showing that we do not support it. There are workarounds available in case this unlikely issue happens in production. If this issue seems to be more frequent, we might just implement the fix.

Reviewed By: andreacampi

Differential Revision: D68279692

fbshipit-source-id: b5088986ceb4e2332adc2e3e37d56b9f7bd4c8f6