Implement support for refs pointing to trees and blobs in Mononoke Git read path
Summary: With the changes in this diff, should be able to support reading (i.e. cloning, fetching) of refs pointing to trees and blobs. However, the current support is insufficient for standalone trees and blobs, i.e. if no other ref or commit points to them. That will be added later in the stack with a dedicated integration test of its own.

Differential Revision: D68271126

fbshipit-source-id: 93dee4427ae2ca7fa723942efec107570a14c0bf