Bump the number of max retries when adding an entry to the bookmarks log
Summary:
For $reasons we don't use an autonumber field, but instead Mononoke tries to predict the next free I and then is supposed to handle the insertion clash with a retry. Even different bookmarks will race for the log.

The existing 5 retries seem to be enough for most services, but async worker and git server are hitting the limit: https://fburl.com/canvas/gapffmw3. These are two services that are likely to process spikes of commits for the same repo.

Doubling to 10 should give us a bit more safety margin. Longer term we might need to redesign.

I tried making this configurable via CLI flag but that is more than I have time for now.

Reviewed By: clara-9

Differential Revision: D68328195

fbshipit-source-id: 2092e4ff1bedb9721405d05b7a9085521e0daac8