Fail import of a repo with tree refs if not run with --allow-content-refs
Summary: For some reason, we bail in case of blob refs but in case of tree refs we silently ignore and continue the gitimport process. This makes it confusing when trying to analyze repos for the presence of refs pointing to trees and blobs. Let's have a consistent behavior with an integration test that would protect against regression.

Reviewed By: andreacampi

Differential Revision: D68325469

fbshipit-source-id: 539a50b63dd982738bea14d98b85ab25695053d4