Update large number formatting in eden stats
Summary:
It's hard to parse the "Unloaded, tracked inodes" number without thousand separator. Also as a follow-up to D63347388 I switched it to always use comma for thousands separator since vast majority of people reading this will likely be in US locale. And I didn't want to add locale code just for this.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: MichaelCuevas

Differential Revision: D68335985

fbshipit-source-id: acef429a4f8f06ff91758d8b938f2e6d03305d66