commit_graph: allow using read-committed transaction level gated behind a JK.
Summary: We've been seeing lock wait timeouts for some time when inserting new changesets into the commit graph. Switching transaction isolation level to read-committed is one the most recommended mitigations for this and the insertion queries do not rely on behaviours of repeatable read so it's worth giving it a try.

Reviewed By: RajivTS, singhsrb

Differential Revision: D68417791

fbshipit-source-id: a77760b519a9db11bbc271a694a1c43bb883c384