modern sync: allow it to override bookmarks
Summary:
When make operations on the bookmarks table we gather operations to execute on a transaction  and classify them on any of the CRUD actions, however, we don't support updating an entry through the creation workflow and we need it for modern sync.

Repos like admin instead of setting `master` to commit A and then move commit A to commit B they simply have a bookmark creation for master to commit A and then same for commit B, so when we were importing admin the bookmark wasn't being updated given we don't support bookmark overrides through edenapi, so adding this now.

So here I'm adding a new reason for bookmark move which can be set by a pushvar when sending requests through edenapi, putting this up for review but I don't intend to land it until I add the ACL checking.

Reviewed By: clara-9

Differential Revision: D68398738

fbshipit-source-id: d44a0e1037f17a75f015407d4ab81974c9200312