Remove unused RefUpdateOperation struct
Summary: This struct was added in `bookmark_mover` in Mononoke Git server to contain information about a ref update operation in addition to the basic `(ref_name, from, to)` triple. However, the current implementation doesn't include any additional context in this struct and is essentially an unnecessary wrapper on RefUpdate itself. Let's get rid of this unnecessary wrapper.

Differential Revision: D68411777

fbshipit-source-id: a24f6b0359c23168867e2dc6466fc4e3d831e84c