Implement forEachDeltaForwards
Summary:
# Context

 forEachDelta is a function that allows users to call a callback on each delta recorded in EdenFS Journal starting from the current delta towards a specified point.

The existing implementation has some issues, such as not being able to cancel iteration based on the current conditions(eg. too many results), and needing to reverse the results at the end for desired order.

# This Diff
This diff implements forEachDeltaForwards

It starts at the end of fileChange/hashUpdate arrays and steps backwards until either
1) It is at the beginning of the array(aka the specified point is before the start of the array)
2) The sequence value of the current item is one item before the specified point.
In the case of 2, we advance once so the sequence number is greater than or equal to the specified point

We then advance forwards to the current delta, calling the appropriate callback along the way. If either callback returns false, we exit early

# Discussion
Should we set  deltaState->lastModificationHasBeenObserved = true when exiting early?

Reviewed By: jdelliot

Differential Revision: D68349278

fbshipit-source-id: 3fe001b598d33ac9e4e14d3e7beb4bba5cd82541