Syntax highlighting in absorb
Summary:
Instead of bespoke line rendering, reuse <SplitDiffTable> to get automatic syntax highlighting.

We just need to construct the appropriate context and Patch.

This gives us intra-line highlighting as well, which is quite nice.

This does "double up" line numbers, which for pure additions shows an extra gap. I think that's fine.

We need to do a little path computation to find the file path to do syntax highlighting when dragging & dropping, but it's basically copy pasta from the non-drag case.

Reviewed By: quark-zju

Differential Revision: D68337944

fbshipit-source-id: 0e5c70cc8311e98a841f8a95344e9652abfc43b3