Add "send to commit" button
Summary:
Add in a button that appears when hovering on an absorb edit. Clicking it shows a context menu with all the valid absorb destinations. Clicking on will send that edit to that commit, as if dragging and dropping.

This provides an alternate UX to drag&drop, which can be difficult for large changes.

Right now, if an edit has no valid absorb destination, this button does nothing, which is bad. We should change this.

Reviewed By: nsblake

Differential Revision: D68469011

fbshipit-source-id: c54cc5ce88abed826a69c8f7db02db45434eeac2