Always include Wdir as a candidate for absorbing
Summary:
The next diff adds a "send directly to commit" dropdown. This involves iterating the valid destinations for a given absorb edit, and presenting them as a context menu.

I noticed Wdir wasn't typically included in the list of locations. I believe this is because the range needs to be inclusive. From the uses of `getAbsorbCommitRevs`, this seems okay to do? Callers could filter out Wdir if they don't want it.

Reviewed By: nsblake

Differential Revision: D68469010

fbshipit-source-id: 0bb2382a85d657164645ca26451370ecedf569e5