Derive Git Commits for Bonsais on demand in Mononoke Git Server
Summary:
We have seen a [recurring issue](https://fburl.com/scuba/mononoke_git_server/wkggbvrz) where `LsRefs` fails due to be missing BonsaiGitMapping for the requested bookmark. This is because our bookmarks are defined in terms of bonsais and when we try to lookup the git commit id corresponding to the bonsai commit id, we fail.

Normally, this should not happen since for every push handled by the Git server, we derive both the bonsai commit and the git commit and store the mapping before accepting the push. However, commits can get landed in this repo bypassing Git server (e.g. SCS). This diff ensures that even in scenarios where bonsai commits get landed without deriving their Git counterparts, the Git server will take care of deriving them on demand.

Reviewed By: andreacampi

Differential Revision: D68492146

fbshipit-source-id: c79c4b43c6eed200a7b10b62e74750c11d629765