Tests for forEachDeltaForwards pt4: only file/hash changes + bugfix for empty delta
Summary:
# Context

 forEachDelta is a function that allows users to call a callback on each delta recorded in EdenFS Journal starting from the current delta towards a specified point.

The existing implementation has some issues, such as not being able to cancel iteration based on the current conditions(eg. too many results), and needing to reverse the results at the end for desired order.

# This Diff
This diff adds tests for  forEachDeltaForwards and a bugfix

This test case covers the following cases:
1) Only file changes
2) Only hash updates

The bugfix covers the case when either fileChangeDeltas or hashUpdateDeltas is empty.
It changes the following behavior
- if delta.end() == delta.begin(), the array is empty. In this case do not decrement the iterator.
- Swap check order of iterator == delta.begin() and dereference for seqId.
- if delta.empty, do not increment after decrementing iterator. Check .empty() directly because end() may == begin() for non-empty iterators

# Discussion

Reviewed By: jdelliot

Differential Revision: D68469772

fbshipit-source-id: 24123164c90d5cea31db025a6de98a89ea428249