Omitting chef last run checks for OD
Summary:
Reason:
Chef does not run on ODs because these machines can be reserved for only 18 hours, and sudo access is not permitted on ODs.
More details in this WP post -
https://fb.workplace.com/groups/vscode.feedback/permalink/9190771900989273/

Background:
Our goal is to ensure that Chef is functioning correctly on the user's machine to confirm they have the latest version of Eden installed. Currently, this check is not limited to any specific platform. In this case, it seems that the health report cannot access the Chef log path to accurately read the last updated timestamp, leading to an exception block being triggered. This check was introduced in the following diff: D66439133.

Reviewed By: genevievehelsel

Differential Revision: D68587542

fbshipit-source-id: 0e8b6288ca0cddc9315d5e6ce61df34ec99f4798