modern sync: re-work bookmark transaction
Summary:
So for force-updating bookmarks I was using a pushvar called `MIRROR_UPLOAD` which basically changed the reason the bookmark was being set in the bookmark update log. Turns out that the `bookmarks_update_log` table has `reason` as an enum, so whenever we inserted something with the `MirrorUpload` use case it didn't actually fill that column and then failed on read.

For fixing, instead of changing the `reason` field, I'm simply  splitting the queues for `create`(normal creation workflow) and `creates_or_updates`(force-set creation workflow) and calling each depending on the pushvar and after checking the ACL.

Reviewed By: andreacampi

Differential Revision: D68678770

fbshipit-source-id: dec54e04d2f11ea93eca8d85eb6b6010602163bb