Export new per-repo queue stats
Summary:
## This stack

We have a recurring issue that causes alerts tailored towards AOSP to trigger, because we don't break them down by repo.

See https://fb.workplace.com/groups/1708850869939124/permalink/1884589562365253/ for context.

## This diff

What it says on the tin.

We need to be careful here: the DB layer only returns results if they are present in the DB. If we applied them naively, we would end up never resetting these counters / gauges when entries for a repo or a status disappear from the DB. This has caused false alerts before.

Reviewed By: singhsrb

Differential Revision: D68628272

fbshipit-source-id: 5b3d8ca31fb2f5ead192c3e227ccdadaf57541b3