Refactor existing stats
Summary:
## This stack

We have a recurring issue that causes alerts tailored towards AOSP to trigger, because we don't break them down by repo.

See https://fb.workplace.com/groups/1708850869939124/permalink/1884589562365253/ for context.

## This diff

Back when I implemented the existing stats, modeling with a hash was expedient, because we could easily enumerate the keys, and it was only 4 of them. Now that we have a better implementation, it's better to stay consistent.

Reviewed By: singhsrb

Differential Revision: D68628270

fbshipit-source-id: d57d107d2d59f0de47c24614769c2687ba646ad1