modern sync: preserve order after lookup
Summary:
opsfiles sync was getting stuck at a bunch of random commits claiming it's parent doesn't exist. This was because when we do a `lookup`, order of responses is not guaranteed, so if I make a `lookup` requests for `[commitA, commitB]`, I could get `[NotPresent(commitB), NotPresent(commitA)]` which is what I was iterating on, hence if `A` is a parent `B`, `B` would fail syncing.

Surprisingly we didn't hit this before, probably something that only manifests when large amounts of commits are queried for

Reviewed By: andreacampi

Differential Revision: D68694024

fbshipit-source-id: 55df94ca0801ba874eee310cb3e4563b2a8882b5