Support --launch in yarn dev
Summary:
Add `--launch <cwd>` flag to `yarn dev` to spawn an ISL server or launch VS Code after doing the client+server build.

This makes it even more convenient to use than before: you only need one terminal window compared to 3.

By default, we don't launch unless you provide the flag. This is because it's often useful to see the full ISL server output or launch vscode via the launch button in the UI instead.

It's a bit annoying to put the isl server `yarn serve` output in the yarn dev output, because we only show ~15 lines of output (depending on your terminal size). But if you want full output, you should probably not use `--launch`, I guess... Maybe ideally you can press a button to dump to a file or open a full screen scrollable `less`-like view of the full server output. I don't bother with that here.

To make this better, in following diffs we will:
- Add quick buttons to restart the server via terminal input
- Support launching internal vscode (requires separate command + flags)

Reviewed By: quark-zju

Differential Revision: D68734890

fbshipit-source-id: ae6644f85134ff6b07acdc8171e528c515ca4dd2