`yarn dev` command to build client & server in parallel
Summary:
Add in a new `yarn dev` command, runnable from `addons/`, which is aimed at simplifying the ISL build process.

Previously, you needed 3 terminal tabs / windows:
- `yarn start` in isl/
- `yarn watch` in isl-server/
- `yarn serve --dev ...` in isl-server/

This new `yarn dev` command is run from `addons/` and runs the first two of these in parallel and shows output in one window.

You still need to launch the server, but it's nice for the build part to be a single command.

In the future, we could make this command optionally launch the server too. (or launch vscode if building for vscode). I don't do that here yet, partly because it's useful to see the full server output instead of a truncated section of it.

I played with implementing this by calling vite and rollup programmatically. It's *dramatically* simpler if we just shell out to `yarn start` and `yarn watch` commands and print the output.

So this diff adds a generic system to run multiple processes and print their stdout in a combined way.

This is inspired by [`bun --filter`](https://bun.sh/blog/bun-v1.2#bun-run-filter). We don't currently use bun so I reimplemented this myself so other people can get the benefit of this system without installing bun and changing our build to support their implementation (it requires code changes in our case).

Reviewed By: quark-zju

Differential Revision: D68531705

fbshipit-source-id: 6ff37e20518ca8aae6bb7db5408c6e5c72fa727e