Support [de]serializing new NFSMountOption fields
Summary:
# This diff

Teaches `[de]serializeNFSMountOptions` to [de]serialize the newest NFSMountOptions. Note: all of these mount options are being added at the same time, so they will either all be present, or all be absent. Therefore the compatibility logic boils down to:

* Are any of the options present?
  * If yes, [de]serialize all of them
  * If not, [de]serialize none of them

The same logic works for deciding whether to use the new options when issuing a mountNFS() request to the privhelper server.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68602728

fbshipit-source-id: 77c58db156c51efb1ff9acbf8db5492e536ccc9c