PrivHelper: use NFSMountOptions for [de]serialization
Summary:
# This diff

Actually uses the new NFSMountOptions struct for serializing/deserializing nfsMount privhelper requests.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68280786

fbshipit-source-id: 9d288a68e84d1d8ecef54a009b5ca43dfb6d6c77