PrivHelper: add new NFSMountOptions
Summary:
# This diff

Adds the newly added EdenConfig options from the last diff to NFSMountOptions. Note: They are all added as optional, so that we can maintain compatibility with old EdenFS Daemons and old privhelpers. Some of the fields are also *double optional*, in the sense that they are not required to be filled out or provided to the mount_nfs command. An example is directory read size, as the "default" option is calculated based on other mount options. Instead of manually calculating what the default should be, we will simply not provide the option and allow the nfs_mount command to choose the default for us.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68602733

fbshipit-source-id: d09a1f330a12f45e381d91fdc0b65d540d41ed4d