PrivHelper: specify additional options for NFS mounts
Summary:
# This diff

Teaches PrivHelperServer to read the new options from the NFSMountOptions struct and pass them into the mount_nfs() invocation accordingly.

NOTE: on Linux, half of these NFSMountOptions are not available. I've modified EdenConfig to indicate which of these options will be no-ops on Linux.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68280782

fbshipit-source-id: f54090f144a7f080ec5104d844765d4ea23a329e