PrivHelper: remove soft mount fallback
Summary:
# This diff

Removes the compatibility code that was added when introducing the hard/soft mount options. It has been 3 months since this code rolled out, so we should no longer need the compatibility fallback.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68280783

fbshipit-source-id: 5baa6697e49eab2c5b9e48745f54d3325d388296