PrivHelper: introduce function for [de]serializing uint8_ts
Summary:
# This diff

Adds [de]serialization functions for `uint8_t` values. While I could refactor `[de]serializeBool` and reuse the logic for both (with an additional cast for bool), I decided the code duplication was minimal.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68602731

fbshipit-source-id: 6aff3f7c1eb53b64d0477c0db093a97933cac28d