PrivHelper: add local locks documentation
Summary:
# This diff

Adds some documentation to the local locks mount option. NOTE: it doesn't make sense to add this as an NFSMountOption, because we don't actually have the ability to change this option. We do not implement locking ourselves, so we must use locallocks exclusively.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68280785

fbshipit-source-id: d19138beb0be6856a3ddb3734d5f8d4839503f9d