PrivHelper: add compatibility fallback for old NFSMountOptions
Summary:
# This diff

Teaches `nfsMount()` to fall back to the old NFSMountOptions if the PrivHelperServer does not understand the request with extra options.

* Did the initial request error?
  * If yes, try passing an NFSMountOptions request with only the original NFSMountOptions
  * If not, no action is needed

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68631084

fbshipit-source-id: d5b891b3ad0ec5d8c8a47a1d1a7bd24e2b16a3a2