PrivHelper: refactor NFS mounting
Summary:
# This diff

refactors the existing nfsMount() code to use the new NFSMountOptions struct that was introduced earlier in the stack. NOTE: This should not require any compatibility logic changes because the same fields are being [de]serialized in both versions of the code. Therefore, this is purely a refactor (and no protocol changes are actually being made).

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68279203

fbshipit-source-id: d548bdad48dfeb8198430ccc726e9d8edd26e31d