PrivHelper: Introduce NFSMountOptions
Summary:
# This diff

Introduces a struct that defines all the options that can be specified when mounting an NFS Mount.

# Context

NFS mounts are extremely configurable, especially on macOS. The configurations vary a ton and impact mount perf/reliability in many different ways. Instead of passing around all these mount options as a list of parameters, I decided to clean up the code and pass them around in a struct. That will keep the nfsMount() signature from growing out of hand, and overall make the code cleaner. This clean up is motivated by a desire to add even more NFS Mount Option configurations in the future (which is done later in this stack).

Reviewed By: jdelliot

Differential Revision: D68280784

fbshipit-source-id: 6c34c800e72738054b1979e198a75c83823406f1