Introduce CommitSyncRepos for transition
Summary:
## This stack
This stack will bring lots of long-planned improvements to the cross-repo codebase.

They're mostly focused on TODOs and known pain points.

The next few diffs will work on migrating `CommitSyncRepos` to a new implementation will make the code much easier to understand.

## This diff

Start the migration of `CommitSyncRepos` to a new implementation that will simplify the abstractions and make the code easier to understand and work with.

This has been a known pain point I discussed with markbt and mitrandir77 over a year ago, but it's a non-trivial refactoring to fix.

## Problem
The cross repo codebase was created a long time ago with very generic abstractions, because it wasn't very clear at the time what exactly we might do later on.

Now that the system is mature and we know exactly what we want to achieve (i.e. merge small repos into fbsource), some of this genericness can make the code a lot more confusing...

The main example is how we refer to repos and synd direction. Right now, each place might refer to repos as small, large, source or target. Some times they need both and throughout the code there is many duplicated logic that basically keeps converting from one format to another...

## Solution
We'll update the data model to stop referencing repos as source and target. We'll always store them as **small or large** and then store the sync direction as well, which will be referred to as **forward or backwards** (instead of SmallToLarge or LargeToSmall). This terminology of forward/back syncing has been used for a long time, but the naming is not used properly in the code.

The first step to do this is to change `CommitSyncRepos`, which is the data type that stores the repos and is used everywhere to access them.

The current implementation is

```
pub enum CommitSyncRepos<R> {
    LargeToSmall {
        large_repo: R,
        small_repo: R,
        submodule_deps: SubmoduleDeps<R>,
    },
    SmallToLarge {
        small_repo: R,
        large_repo: R,
        submodule_deps: SubmoduleDeps<R>,
    },
}
```

You can see the new one below, where we'll use a struct instead of an enum with duplicated fields.

## How we'll implement this change?
1. Introduce the new data type, **temporarily** called `CommitSyncReposWithDirection`. This will come with convenience methods for conversion between the new and old.
2. I'll slowly replace the references to start using the new one and will use the conversion methods in the intermediate layers.
3. After all references are replaced, I'll delete the old implementation and will rename the new one to `CommitSyncRepos`.

There will also be (probably many) other changes to improve this code, e.g. improve the method names, rename the `CommitSyncDirection` type to forward, backwards, etc..

## Notes
- All associated functions are unused in this diff, but they'll start to be used in the following ones. You can see that they're mostly copied from the `CommitSyncRepos`.

Reviewed By: lmvasquezg

Differential Revision: D68555002

fbshipit-source-id: 61ceea8bbd84f18a29b482bb63dfa4613f251ce6