Delete instantiantions of CommitSyncRepos using source and target repos
Summary:
## This stack
Please see the summary of D68555002 for details.

## This diff
We should attempt to reference repos as small/large **when we already have that information**, instead of using source/target.

This diff replaces all the calls to `new_with_source_target` and all calls to `from_source_and_target_repos` **that we can remove** to use the **new** constructor.

## Notes
- The naming is very confusing, so in the next diff D68706369 I do a batch renaming of variables to use small/large terminology as well.
- In some callsites we still need to use `from_source_and_target_repos`, which is why I didn't delete it. These are the callsites that do now know the direction of the sync (i.e. forward/backwards), they just know the source and target repos.

Reviewed By: lmvasquezg

Differential Revision: D68705521

fbshipit-source-id: b9bb9b862c2c9bc53273dfd4eecd34b1c585b741