buffer: avoid aliasing MmapMut references
`Buffer` has an `UnsafeCell<MmapMut>`. It protects against concurrent
writes to it by using a mutex. However, it has UB by allowing callers
of `read()` and `write()` to concurrently have mutable and immutable
references to the `MmapMut`. This patch avoids the alias by using a
pointer instead.

Thanks to @Ralith for spotting the problem and suggesting the fix
during our review of unsafe Rust code at Google.