teach eden reloadconfig to handle --manifold-key
Summary:
# Context
To test edenfs_config changes, one can upload the raw config to manifold and ask the edenfs_config_manager to use that raw config rather than canarying the change to a machine or having a local configerator checkout.

# This Diff
I added this flag mid last year (D61072490) but didn't add it to the helper cli `reloadconfig`. This diff adds it to `reloadconfig` for ease of use. `reloadconfig` just passes the args to the `edenfs_config_manager` so you don't have to remember the path for invoking the manager.

Reviewed By: ViniGupta08

Differential Revision: D68801571

fbshipit-source-id: 8ae36a96886d6a50a108f158a254a7ecfbe381f8