Explicitly kill process children on exit
Summary:
when testing `yarn dev` I noticed sometimes `vite` processes would be left dangling, which would cause future `yarn dev` invocations to not run on port 3000 (it searchs up for an unused one), which breaks the main server command.

It seems we need to explicitly kill the subprocesses we spawned on exit. I was under the impression this was not necessary unless the process we spawn uses `detatched` (which we do not), but I still noticed this behavior in practice. Killing the children seems to fix this.

Reviewed By: nsblake

Differential Revision: D68807606

fbshipit-source-id: 5fa30f5272512b93b306d2c51480c4bc43c7e445