Listen for user input to restart server
Summary:
When using `--launch`, it's useful to restart the server process but leave the builds running.

Add a user input listener, so you can press "R" to restart the running server, or "q" / ctrl-c to quit.

We only do this if using `yarn dev browser --launch`. Doesn't really help for vscode (just reload the window instead), and don't need this if not launching.

Reviewed By: quark-zju

Differential Revision: D68807038

fbshipit-source-id: 9d6596b206ef535d02bbf3b390dddf03bde26df2