prefetch: ignore file paths
Summary:
When fetching file contents, the file paths are used mainly for informational purposes. "prefetch" fetches a lot of files, so the file paths easily add up to a few GB of memory usage. I think it's worth omitting the paths for "prefetch" to save memory.

The downside is that file paths don't show up in certain log messages, and any future smarts that rely on file paths won't work for "prefetch".

Reviewed By: sggutier

Differential Revision: D68355874

fbshipit-source-id: 27867ee532e1806c6476a906ed59b11889e63035