pyrevisionstore: kill use of ContentDataStore trait
Summary:
The only live use was filescmstore.metadata() -> ... -> (scmstore::FileStore as ContentDataStore).metadata(). Kill the stuff in the middle, and the unused "metadata()" method for the treescmstore.

This method currently only returns metadata for LFS pointers already available locally. It serves two purposes. First it acts as aux data for LFS files to optimize some code paths (this can be replaced by real aux data). Second it is used during wireproto push to know a file is LFS (so the bundle should contain a pointer, not the file contents). This can go away once we remove bundle push support (in favor of SLAPI push).

Reviewed By: sggutier

Differential Revision: D68522459

fbshipit-source-id: 4f606443b24acc78683396921c623f1578d4cdcf