scmstore: don't store file/tree path in indexedlogdatastore
Summary:
Ignore the path when reading from the indexedlogs, and write a zero length path for new entries. This saves space on disk and allocations copying the paths everywhere.

When storing file and tree content in indexedlogdatastore, we previously would also store the file/tree's path (e.g. "foo/bar.c"). This is not necessary since the stores are keyed only by the node (AKA hgid).

The downsides to storing the path are:
1. Takes up extra space on disk.
2. Requires passing/cloning the `Key` struct (including the path) a lot.
3. Is not actually reliable since two different files/trees can share exactly the same node. The code was aware of this since it would call `with_key(key)` to override the stored path with the in-hand path (since they could be different).

The advantage is you can know the name of a file/tree when inspecting/iterating the indexedlog stores.

I've never needed to inspect the path, so I feel the downsides outweigh the advantage.

Also note that files and trees fetched via EdenFS already never have the path set (because EdenFS leaves the `Key` path empty).

Reviewed By: liubov-dmitrieva

Differential Revision: D68522462

fbshipit-source-id: 62bf2b1de4620ee1e3c062bd6ccde99c89a59c46