Remove dependency on `failure_ext` crate ahead of migrating from slog to tracing.
Summary:
This crate used to be a wrapper around `failure`, before `anyhow` became the de-factor standard and made it unnecessary.

It is now a wrapper around `anyhow` that doesn't add much to the functionality from anyshow.

In addition, it is coupled with `slog` which is getting deprecated in favour of `tracing`.

In all call-sites under `eden/mononoke`, replace `failure_ext` idioms with more standard constructs.

In particular, replace both
* [`DisplayChain`'s `Display` implementation](https://www.internalfb.com/code/fbsource/[ea26abb3cde106c14d25d0e640f4c42ec36dba92]/fbcode/common/rust/shed/failure_ext/src/lib.rs?lines=61) and
* [`SlogKVError`'s `serialize` implementation](https://www.internalfb.com/code/fbsource/[ea26abb3cde106c14d25d0e640f4c42ec36dba92]/fbcode/common/rust/shed/failure_ext/src/slogkv.rs?lines=22)

with straightforward uses of the `Debug` trait on `anymow::Error`, which conceptually does the same thing:
display the details of the error in a human readable way.

Reviewed By: markbt

Differential Revision: D68835972

fbshipit-source-id: 96ed3750e00dde4b8ecee6605e45ba73fbc6883e